fix(build): guard Linux-only inotify/signalfd headers so macOS/iOS compile
- <sys/inotify.h> and <sys/signalfd.h> only exist on Linux; the unconditional
  includes broke the macOS arm64 build. Guard them with #if defined(__linux__)
  — Apple builds already use the loopback-UDP emulation branch
- elf_loader.cpp: fix -Wformat warnings in relocate summary (%zu vs uint64_t)
  by using %llu with explicit casts

+++ src/elf_loader.cpp
@@ -701,11 +701,11 @@ bool ElfLoader::relocate() {
                     applyRelocations(jmpRelVaddr, jmpRelSize) &&
                     applyRelr();
     fprintf(stderr,
-            "[KuDroidELF] relocate(%s): RELA=%zu entries, JMPREL=%zu, RELR=%zu bytes, symtab=%s\n",
+            "[KuDroidELF] relocate(%s): RELA=%llu entries, JMPREL=%llu, RELR=%llu bytes, symtab=%s\n",
             path_.c_str(),
-            relaSize / relaEnt,
-            relaEnt ? jmpRelSize / relaEnt : 0,
-            (size_t)relrSize,
+            static_cast<unsigned long long>(relaSize / relaEnt),
+            static_cast<unsigned long long>(relaEnt ? jmpRelSize / relaEnt : 0),
+            static_cast<unsigned long long>(relrSize),
             symtabVaddr != 0 ? "present" : "absent");
 #if defined(__APPLE__) && TARGET_OS_OSX
     pthread_jit_write_protect_np(1);
